requirements.txt
fastapi[standard]
uvicorn
langchain
langchain-community
langchain-google-genai
langchain-openai
pymongo
passlib[bcrypt]
pyjwt
openai
langgraph